requirements.txt
amqp==2.5.0
appnope==0.1.0
backcall==0.1.0
billiard==3.6.0.0
celery==4.3.0
certifi==2019.6.16
chardet==3.0.4
decorator==4.4.0
dj-database-url==0.5.0
Django==2.2.3
django-heroku==0.3.1
gunicorn==19.9.0
idna==2.8
ipdb==0.12
ipython==7.6.1
ipython-genutils==0.2.0
jedi==0.14.1
kombu==4.6.3
parso==0.5.1
pexpect==4.7.0
pickleshare==0.7.5
Pillow==6.1.0
prompt-toolkit==2.0.9
psycopg2==2.8.3
ptyprocess==0.6.0
Pygments==2.4.2
pytz==2019.1
redis==3.2.1
requests==2.22.0
requests-toolbelt==0.9.1
six==1.12.0
sqlparse==0.3.0
traitlets==4.3.2
urllib3==1.25.3
vine==1.3.0
wcwidth==0.1.7
whitenoise==4.1.3
